+++ src/core/EntityID.ts
@@ -1,10 +1,12 @@
 import crypto from 'node:crypto';
 
+import type { IComponent } from '@/core/types/index.js';
+
 /**
  * @file Manages entity identifier.
  * @copyright Piggly Lab 2023
  */
-export class EntityID<Value = string> {
+export class EntityID<Value = string> implements IComponent {
 	/**
 	 * Indicates if identifier was randomly generated.
 	 *
@@ -75,6 +77,20 @@ export class EntityID<Value = string> {
 		}
 
 		return id.value === this.value;
+	}
+
+	/**
+	 * Checks if the object is a specific component type.
+	 *
+	 * @param {string} name
+	 * @returns {boolean}
+	 * @public
+	 * @memberof Entity
+	 * @since 5.0.0
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	public is(name: string): boolean {
+		return name.toLowerCase() === 'entityid';
 	}
 
 	/**